requirements.txt
Pillow
mysqlclient
redis
hiredis
gevent
oss2
Django
django-redis-sessions
django-redis-cache
djangorestframework
django-restframework-apiview>=1.2.20
celery
pycrypto
dingtalk-sdk>=1.1.0
